+++ distnet/__main__.py
@@ -12,9 +12,9 @@ sys.path.append(os.path.realpath(os.path.dirname(__file__)+"/.."))
 sys.path.append(os.path.realpath(os.path.dirname(__file__)+"/../distnet"))
 
 from ioctl.io_controller import IO_Ctl
-from network_controller import resolve_hostname
-from tcp_file_handler import FileTCP
-from os_check import os_filesystem_check
+from networking.network_controller import resolve_hostname
+from tcp.tcp_file_handler import FileTCP
+from utils.os_check import os_filesystem_check
 
 
 import pkg_resources
