+++ tests/test_network_controller.py
@@ -5,7 +5,7 @@ import pytest
 sys.path.append(os.path.realpath(os.path.dirname(__file__)+"/.."))
 sys.path.append(os.path.realpath(os.path.dirname(__file__)+"/../distnet"))
 
-from network_controller import resolve_hostname, resolve_location
+from networking.network_controller import resolve_hostname, resolve_location
 
 def test_invalid_resolve_hostname():
     with pytest.raises(Exception) as exception_info:
@@ -19,4 +19,4 @@ def test_resolve_location_tldstring():
     assert resolve_location('whatastrangehostname.uk') == 'uk'
 
 def test_resolve_location_us():
-    assert resolve_location('google.com') in ['us', 'com'] 
+    assert resolve_location('google.com') in ['us', 'com']
