+++ tests/test_os_check.py
@@ -6,7 +6,7 @@ import pytest
 sys.path.append(os.path.realpath(os.path.dirname(__file__)+"/.."))
 sys.path.append(os.path.realpath(os.path.dirname(__file__)+"/../distnet"))
 
-from os_check import os_filesystem_check
+from utils.os_check import os_filesystem_check
 
 def test_os_filesystem_check_empty():
     try:
