+++ tests/test_tcp_structs.py
@@ -5,9 +5,9 @@ import sys
 sys.path.insert(0, os.path.abspath('..'))
 sys.path.insert(0, os.path.abspath('../distnet'))
 
-from tcp_structs import C_STATE
-from tcp_file_entry import EntryTCP
-from hex_manip import int_from_string
+from tcp.tcp_structs import C_STATE
+from tcp.tcp_file_entry import EntryTCP
+from utils.hex_manip import int_from_string
 
 
 def test_c_state():
